+++ stochproc/timeseries/ssm.py
@@ -40,7 +40,7 @@ class StateSpaceModel(StructuralStochasticProcess):
         self._dist_builder = f
 
         self.observe_every_step = observe_every_step
-        self._event_shape = self.initial_sample()["y"].event_dim
+        self._event_shape = self.initial_sample()["y"].event_shape
 
     def initial_dist(self) -> Distribution:
         raise NotImplementedError("Cannot sample from initial distribution of SSM directly!")
@@ -49,7 +49,7 @@ class StateSpaceModel(StructuralStochasticProcess):
         init_dist = self.build_density(x)
         empty = float("nan") * torch.ones(init_dist.batch_shape + init_dist.event_shape, device=x.values.device)
 
-        return StateSpaceModelState(x=x, y=TimeseriesState(x.time_index, values=empty, event_dim=init_dist.event_shape))
+        return StateSpaceModelState(x=x, y=TimeseriesState(x.time_index, values=empty, event_shape=init_dist.event_shape))
 
     def initial_sample(self, shape: torch.Size = torch.Size([])) -> StateSpaceModelState:
         x_state = self.hidden.initial_sample(shape)