+++ tests/distributions/test_module.py
@@ -1,7 +1,7 @@
 import torch
 
 from stochproc.distributions import DistributionModule
-from pyro.distributions import Normal
+from pyro.distributions import Normal, Independent
 
 
 class TestDistributionModule(object):
@@ -21,3 +21,8 @@ class TestDistributionModule(object):
 
         dist = mod.build_distribution()
         assert dist.mean.device == torch.device("cuda:0")
+
+    def test_expand(self):
+        mod = DistributionModule(Normal, loc=0.0, scale=1.0).expand(torch.Size([2])).to_event(1)
+        dist = mod.build_distribution()
+        assert isinstance(dist, Independent) and isinstance(dist.base_dist, Normal)