+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Versions
 
+### v0.0.14
+ - Adds the `batch_shape` property to `TimeseriesState`.
+ - Renames `event_dim` property of `TimeseriesState` to `event_shape`.
+ - Adds support for `expand` and `to_event` on `DistributionModule`.
+
 ### v0.0.13
  - Adds the `SmoothLinearTrend` model.